fix: use split instead of regex to extract metric group
This was flagged by sonarqube under python:S5852 which relates
to catastrophic backtracking. This isn't really relevant here
tbh because this code is in the cms admin view and therefore not
going to be under any kind of load, however, might as well fix
it rather than have to add a `# noqa` of some variety. This code
now matches the validations way of handling this too so that's
no bad thing.